build/ASL: remove "copy after compile" step
The files `dist/src/stepFunctions/asl/aslServer.js` and
`dist/src/stepFunctions/asl/aslServer.js.map` already are created by the
compile step. Why was `filemanager-webpack-plugin` ever needed?

- Remove `copy-webpack-plugin`
- `copy-webpack-plugin` doesn't have an `onEnd` concept like
  `filemanager-webpack-plugin` did. Instead we can move the copy step to
  the same place used by SSM. But ASL doesn't actually *need* a copy
  step (see above)...
- where does the *.LICENSE.txt.js file come from?

Test cases:
-----------
Using the `.vsix` artifact, observe:
- `ps aux` shows that `aslServer` is running:
   ```
   jmkeyes 87125   0.0  0.3  8718916  44536   ??  S     8:09AM   0:00.36
   /…/Visual Studio Code.app/ …/.vscode/extensions/amazonwebservices.aws-toolkit-vscode-1.16.0-SNAPSHOT/dist/src/stepFunctions/asl/aslServer.js --node-ipc --clientProcessId=85703
   ```
- Code-completion works on `HelloWorld.asl.json`.
- "Render graph" feature works on `HelloWorld.asl.json`.
