+++ package.json
@@ -1373,8 +1373,8 @@
         "reset": "node ./build-scripts/clean.js dist node_modules && npm install",
         "copyNonCodeFiles": "node ./build-scripts/copyNonCodeFiles.js",
         "copyExtensionMain": "node ./build-scripts/copyExtensionMain.js",
-        "copySsmLanguageService": "node ./build-scripts/copySsmLanguageService.js",
-        "buildScripts": "npm run bundleDeps && npm run copyExtensionMain && npm run copySsmLanguageService && npm run copyNonCodeFiles && npm run generateNonCodeFiles",
+        "copyDistFiles": "ts-node ./build-scripts/copyDistFiles.ts",
+        "buildScripts": "npm run bundleDeps && npm run copyExtensionMain && npm run copyDistFiles && npm run copyNonCodeFiles && npm run generateNonCodeFiles",
         "compile": "webpack --mode development && npm run buildScripts",
         "recompile": "npm run clean && npm run compile",
         "watch": "npm run buildScripts && tsc -watch -p ./",
@@ -1420,7 +1420,6 @@
         "@typescript-eslint/eslint-plugin-tslint": "^2.27.0",
         "@typescript-eslint/parser": "^2.27.0",
         "circular-dependency-plugin": "^5.2.0",
-        "copy-webpack-plugin": "^6.2.0",
         "decache": "^4.4.0",
         "eslint": "^6.8.0",
         "eslint-config-prettier": "^6.10.1",
