+++ webpack.config.js
@@ -11,7 +11,6 @@ const path = require('path')
 const webpack = require('webpack')
 const fs = require('fs')
 const TerserPlugin = require('terser-webpack-plugin')
-const CopyPlugin = require('copy-webpack-plugin')
 const { NLSBundlePlugin } = require('vscode-nls-dev/lib/webpack-bundler')
 const CircularDependencyPlugin = require('circular-dependency-plugin')
 const packageJsonFile = path.join(__dirname, 'package.json')
@@ -23,7 +22,7 @@ const config = {
     target: 'node',
     entry: {
         extension: './src/extension.ts',
-        aslServer: './src/stepFunctions/asl/aslServer.ts',
+        'src/stepFunctions/asl/aslServer': './src/stepFunctions/asl/aslServer.ts',
     },
     output: {
         path: path.resolve(__dirname, 'dist'),
@@ -74,23 +73,6 @@ const config = {
         new NLSBundlePlugin(packageId),
         new webpack.DefinePlugin({
             PLUGINVERSION: JSON.stringify(packageJson.version),
-        }),
-        // @ts-ignore
-        new CopyPlugin({
-            patterns: [
-                {
-                    from: path.resolve(__dirname, 'dist/aslServer.js'),
-                    to: path.resolve(__dirname, 'dist/src/stepFunctions/asl/aslServer.js'),
-                },
-                {
-                    from: path.resolve(__dirname, 'dist/aslServer.js.LICENSE.txt'),
-                    to: path.resolve(__dirname, 'dist/src/stepFunctions/asl/aslServer.js.LICENSE.txt'),
-                },
-                {
-                    from: path.resolve(__dirname, 'dist/aslServer.js.map'),
-                    to: path.resolve(__dirname, 'dist/src/stepFunctions/asl/aslServer.js.map'),
-                },
-            ],
         }),
         new CircularDependencyPlugin({
             exclude: /node_modules/,
